Boiler plate code for taking a penalty function approach to diversifying designs as a replacement for current heuristics.
Commented out the writing to file that happening during running as this was taking over half of the time.

Increased number of samples used in design optimization by factor of 10 as the estimates were poor and we
were spending a lot less time here than in other areas like the inference.