Refined heuristic for spreading out designs.  Now uses a kernel density estimator on the whole space rather than a product of estimators on each dimension separately.
There could still be issues in very misspecified models that it will never force a dimension to take a
value that it has never taken before - the heuristic prevents too many designs being close to each other,
but it does not necessarily force all regions where we would like designs to have them.  This is because
it is complicated to distinguish betweens regions we have not sampled because they genuinely provide
little information and regions we have not sampled because of model mispecification issues / choosing
diverse options between similarly good options.  There is thus scope for further improvements here, such
as using multiple length scales in the kernel or adding additional heuristics to ensure that each
individual design variable has a good spread.